extract state creation and update login in outer class
